requirements.txt
streamlit
streamlit-option-menu
pymongo
langchain-mongodb
openai
python-dotenv
langchain
langchain_openai
langchain-core
langchain_community
pypdf

